cover the different kinds of setup
